refactor:  home controller 이동 및 logout post -> get으로 변경
