+++ .gitignore
@@ -36,3 +36,5 @@ out/
 ### VS Code ###
 .vscode/
 src/main/resources/application-secret.yml
+
+src/main/generated
